+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to the "cursor-feedback" extension will be documented in thi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [1.0.1]
+
+- Using only the VSCode Extension API
+  - no longer using apple script, hence
+    - safer
+    - no longer macOS-specific
+- Custom IDE command trigger
+- Automatic management of started command-line process
+
+## [1.0.0]
 
 - Initial release